fix clang-tidy v5

+++ tasks/mpi/kalinin_d_vector_dot_product/include/ops_mpi.hpp
@@ -36,7 +36,7 @@ class TestMPITaskParallel : public ppc::core::Task {
  private:
   std::vector<std::vector<int>> input_;
   std::vector<int> local_input1_, local_input2_;
-  std::vector<unsigned int> counts_;
+  std::vector<size_t> counts_;
   size_t num_processes_ = 0;
   int res_{};
   boost::mpi::communicator world_;
